+++ src/main/java/org/folio/cataloging/business/cataloguing/bibliographic/PublisherTag.java
@@ -19,6 +19,7 @@ import org.folio.cataloging.dao.persistence.BibliographicNoteType;
 import org.folio.cataloging.dao.persistence.PUBL_HDG;
 import org.folio.cataloging.dao.persistence.REF;
 import org.folio.cataloging.model.Subfield;
+import org.folio.cataloging.shared.CorrelationValues;
 import org.folio.cataloging.util.StringText;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -56,7 +57,7 @@ public class PublisherTag extends VariableField implements PersistentObjectWithV
 	 * noteType is used only for the edit page correlation to allow for the possibility
 	 * that a publisher tag can be changed to another note type.
 	 */
-	private short noteType = 24;
+	private int noteType = 24;
 
 	/**
 	 * Class constructor
@@ -208,7 +209,7 @@ public class PublisherTag extends VariableField implements PersistentObjectWithV
 	/* (non-Javadoc)
 	 * @see librisuite.business.cataloguing.bibliographic.Tag#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 7;
 	}
